network: move neighbor from node to network
